setindex! deprecation phase
